implemented a really hacky approach to #5
so... this works, but I have to admit this project has gotten
pretty out of hand at this point. it's just hack upon hack, with
very little regard for overall quality or even a generally sound
design. so I'm going to push, and then take a step back and
rethink.