strip out leading 'var' from "actual" expressions
This is maybe a bad idea? The deeper you go treating code as mere text,
and mutate it the way you would do string manipulation, the further you
risk running into esoteric bugs due to the impedence mismatch.

Well, for now I'm going with the hack. When it bites me (or sooner) I
will think about a better approach.